更新

+++ project-chashi/src/main/java/com/io/yy/merchant/controller/CsMerchantOrderController.java
@@ -161,7 +161,7 @@ public class CsMerchantOrderController extends BaseController {
      */
     @PostMapping("/getOrderByCurrent")
     @ApiOperation(value = " 根据tearoomid和当前时间判断是否存在订单，有订单返回订单id，没有订单返回null", notes = "判断当前时间是否有订单", response = ApiResult.class)
-    public ApiResult<Integer> getOrderByCurrent(@Valid @RequestBody CsMerchantOrderQueryParam csMerchantOrderQueryParam) throws Exception {
+    public ApiResult<Long> getOrderByCurrent(@Valid @RequestBody CsMerchantOrderQueryParam csMerchantOrderQueryParam) throws Exception {
         return ApiResult.ok(csMerchantOrderService.getOrderByCurrent(csMerchantOrderQueryParam));
     }
