+++ project-chashi/src/main/java/com/io/yy/merchant/service/CsMerchantOrderService.java
@@ -98,7 +98,7 @@ public interface CsMerchantOrderService extends BaseService<CsMerchantOrder> {
      * @param csMerchantOrderQueryParam
      * @return
      */
-    Integer getOrderByCurrent(@Param("param") CsMerchantOrderQueryParam csMerchantOrderQueryParam);
+    Long getOrderByCurrent(@Param("param") CsMerchantOrderQueryParam csMerchantOrderQueryParam);
 
     /**
      * 更新支付状态
